+++ A319 Mr Sweepy/Assets/Sweeper Boi/Material/OptimizedShaders/New Material/Includes/CGI_PoiEmission.cginc
@@ -168,20 +168,20 @@ float3 calculateEmissionNew(in float3 baseColor, inout float4 finalColor)
         }
     #endif
     
-    float glowInTheDarkMultiplier0 = calculateGlowInTheDark(float(0), float(1), float(1), float(0), float(1));
+    float glowInTheDarkMultiplier0 = calculateGlowInTheDark(float(0), float(1), float(1), float(1), float(1));
     
     #if defined(PROP_EMISSIONMAP) || !defined(OPTIMIZER_ENABLED)
         
         if (!float(0))
         {
-            emissionColor0 = POI2D_SAMPLER_PAN(_EmissionMap, _MainTex, poiMesh.uv[float(0)], float4(0,0,0,0)).rgb * lerp(1, baseColor, float(0)).rgb * float4(2,2,2,1).rgb;
+            emissionColor0 = POI2D_SAMPLER_PAN(_EmissionMap, _MainTex, poiMesh.uv[float(0)], float4(0,0,0,0)).rgb * lerp(1, baseColor, float(1)).rgb * float4(2,2,2,1).rgb;
         }
         else
         {
-            emissionColor0 = UNITY_SAMPLE_TEX2D_SAMPLER(_EmissionMap, _MainTex, ((.5 + poiLight.nDotV * .5) * float4(1,1,0,0).xy) + _Time.x * float(5)).rgb * lerp(1, baseColor, float(0)).rgb * float4(2,2,2,1).rgb;
+            emissionColor0 = UNITY_SAMPLE_TEX2D_SAMPLER(_EmissionMap, _MainTex, ((.5 + poiLight.nDotV * .5) * float4(1,1,0,0).xy) + _Time.x * float(5)).rgb * lerp(1, baseColor, float(1)).rgb * float4(2,2,2,1).rgb;
         }
     #else
-        emissionColor0 = lerp(1, baseColor, float(0)).rgb * float4(2,2,2,1).rgb;
+        emissionColor0 = lerp(1, baseColor, float(1)).rgb * float4(2,2,2,1).rgb;
     #endif
